requirements.txt
asgiref==3.2.10
Django==3.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
mysqlclient==2.0.1
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
